🚨 Restored all CLA rules
CLA(S) are still using the "Faculty Resources" section of the old `cla.mercer.edu` site. That causes the large number of rules because they have to work around that fact. That's why DNS was never moved from CommonSpot and cannot be moved from CommonSpot while that section of the site is still active.